requirements.txt
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pycodestyle==2.6.0
python-dotenv==0.15.0
python-sonarqube-api==1.1.9
requests==2.25.1
toml==0.10.2
urllib3==1.26.2
